modified main to include boil_detector class, split camera capture into distorted vs undistorted
